requirements.txt
httpcore==0.16.3
googletrans==4.0.0rc1
kivy
gtts
pygame
speechrecognition
requests
beautifulsoup4
sounddevice
twilio
fpdf
pywinauto
cython
buildozer
kivy-ios
pillow
pyttsx3
pyaudio
numpy
matplotlib
scipy
pandas
openpyxl
xlrd
pyinstaller
pytesseract
keyboard
pyautogui
pygetwindow
pyperclip
# Platform-specific dependencies (install as needed)
# Windows: pywinauto (already included above)
# macOS: pyobjc-framework-Cocoa, pyobjc-framework-Quartz
# Linux: python-xlib, xdotool (system package)
# Android: python-for-android, jnius
# iOS: pyobjc-framework-UIKit, pyobjc-framework-AVFoundation
